fixed libft.h according to linux manual

+++ ft_strlen.c
@@ -6,13 +6,15 @@
 /*   By: rfoo <rfoo@student.42singapore.sg>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2025/11/22 21:30:19 by rfoo              #+#    #+#             */
-/*   Updated: 2025/11/22 21:32:27 by rfoo             ###   ########.fr       */
+/*   Updated: 2025/11/22 22:27:49 by rfoo             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(char *str)
+#include "libft.h"
+
+size_t	ft_strlen(char *str)
 {
-	int	length;
+	size_t	length;
 
 	length = 0;
 	while (str[length])
